openthread: zigbee: lib: Extract ram power down function for common use
To minimize power consumption thread needs to be able to disable unused
ram regions. After some investigation such function was found in zigbee
helpers.
Moved this function to common libraries and made use of it in thread
coap client sample.
