+++ lib/CMakeLists.txt
@@ -20,6 +20,7 @@ add_subdirectory_ifdef(CONFIG_ADP536X adp536x)
 add_subdirectory_ifdef(CONFIG_ST25R3911B_LIB st25r3911b)
 add_subdirectory_ifdef(CONFIG_FPROTECT fprotect)
 add_subdirectory(flash_patch)
+add_subdirectory_ifdef(CONFIG_RAM_POWER_DOWN_LIBRARY ram_pwrdn)
 add_subdirectory_ifdef(CONFIG_RESET_ON_FATAL_ERROR fatal_error)
 add_subdirectory_ifdef(CONFIG_MODEM_KEY_MGMT modem_key_mgmt)
 add_subdirectory_ifdef(CONFIG_SMS sms)
