+++ lib/ram_pwrdn/CMakeLists.txt
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2020 Nordic Semiconductor
+#
+# SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
+#
+
+zephyr_library()
+zephyr_library_sources(ram_pwrdn.c)
