+++ samples/openthread/coap_client/overlay-mtd.conf
@@ -8,3 +8,4 @@
 CONFIG_OPENTHREAD_MTD=y
 CONFIG_OPENTHREAD_MTD_SED=y
 CONFIG_OPENTHREAD_POLL_PERIOD=3000
+CONFIG_RAM_POWER_DOWN_LIBRARY=y
