+++ samples/zigbee/light_switch/prj.conf
@@ -37,3 +37,6 @@ CONFIG_CRYPTO_NRF_ECB=y
 
 # Cryptocell is not supported through CSPRNG driver API: NCSDK-4813
 CONFIG_ENTROPY_CC310=n
+
+# Enable API for powering down unused RAM parts
+CONFIG_RAM_POWER_DOWN_LIBRARY=y
